CVC-909 Nonce Manager singleton
